Pokemon Battle
A little project I did back in school in 2014. It's a partial simulation
of the Pokemon battle system in an arcade game fashion. It was made
using pygame. I have to admit, there is A LOT to be improved with the
structure of the code but, it's still readable, provided you're familiar
with the underlying Pokemon system.